+++ apps/shared/avifpng.h
@@ -7,7 +7,7 @@
 #include "avif/avif.h"
 
 // if (requestedDepth == 0), do best-fit
-avifBool avifPNGRead(avifImage * avif, const char * inputFilename, avifPixelFormat requestedFormat, int requestedDepth);
-avifBool avifPNGWrite(avifImage * avif, const char * outputFilename, int requestedDepth);
+avifBool avifPNGRead(avifImage * avif, const char * inputFilename, avifPixelFormat requestedFormat, uint32_t requestedDepth, uint32_t * outPNGDepth);
+avifBool avifPNGWrite(avifImage * avif, const char * outputFilename, uint32_t requestedDepth);
 
 #endif // ifndef LIBAVIF_APPS_SHARED_AVIFPNG_H